fix: refactored routing to use controllers
