+++ server/app.js
@@ -1,26 +1,26 @@
-const createError = require('http-errors');
-const express = require('express');
-const path = require('path');
-const cookieParser = require('cookie-parser');
-const logger = require('morgan');
-const cors = require('cors');
+const createError = require("http-errors");
+const express = require("express");
+const path = require("path");
+const cookieParser = require("cookie-parser");
+const logger = require("morgan");
+const cors = require("cors");
 const passport = require("passport");
 
-const users = require("./routes/users");
-const menuRouter = require('./routes/menu');
-const db = require('./db/app.js');
+const users = require("./routes/user-router");
+const menuRouter = require("./routes/menu-router");
+const db = require("./db/app.js");
 const app = express();
-const apiPort = 9000
+const apiPort = 9000;
 
 // view engine setup
-app.set('views', path.join(__dirname, 'views'));
-app.set('view engine', 'jade');
+app.set("views", path.join(__dirname, "views"));
+app.set("view engine", "jade");
 
-app.use(logger('dev'));
+app.use(logger("dev"));
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 app.use(cookieParser());
-app.use(express.static(path.join(__dirname, 'public')));
+app.use(express.static(path.join(__dirname, "public")));
 app.use(cors());
 
 // Passport middleware
@@ -30,7 +30,7 @@ require("./config/passport")(passport);
 
 // Routes
 app.use("/api/users", users);
-app.use('/', menuRouter);
+app.use("/", menuRouter);
 
 // catch 404 and forward to error handler
 app.use(function(req, res, next) {
@@ -41,17 +41,17 @@ app.use(function(req, res, next) {
 app.use(function(err, req, res, next) {
   // set locals, only providing error in development
   res.locals.message = err.message;
-  res.locals.error = req.app.get('env') === 'development' ? err : {};
+  res.locals.error = req.app.get("env") === "development" ? err : {};
 
   // render the error page
   res.status(err.status || 500);
-  res.render('error');
+  res.render("error");
 });
 
-app.get('/', (req, res) => {
-  res.send('Hello World!')
-})
+app.get("/", (req, res) => {
+  res.send("Hello World!");
+});
 
-app.listen(apiPort, () => console.log(`Server running on port ${apiPort}`))
+app.listen(apiPort, () => console.log(`Server running on port ${apiPort}`));
 
 module.exports = app;
