+++ server/controllers/menu-ctrl.js
@@ -1,0 +1,1 @@
+// TODO: refactor routes/menu-router.js
