+++ server/routes/user-router.js
@@ -1,0 +1,10 @@
+const express = require("express");
+
+const UserCtrl = require("../controllers/user-ctrl");
+
+const router = express.Router();
+
+router.post("/register", UserCtrl.registerUser);
+router.post("/login", UserCtrl.loginUser);
+
+module.exports = router;
